refactor: json schema
