+++ pkgs/zss/jsonschema/internal/schema.go
@@ -1,0 +1,3 @@
+package internal
+
+type Schema map[string]any
